# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  domain.WishlistItem:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_reserved:
        type: boolean
      name:
        type: string
      priority:
        type: integer
      product_link:
        type: string
      reserved_at:
        type: string
      wishlist_id:
        type: string
    type: object
...
  dto.CreateWishlistItemRequest:
    properties:
      description:
        type: string
      name:
        type: string
      priority:
        type: integer
      product_link:
        type: string
    type: object
...
  dto.UpdateWishlistItemRequest:
    properties:
      description:
        type: string
      name:
        type: string
      priority:
        type: integer
      product_link:
        type: string
    type: object
...
  response.ErrorPayload:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: 'REST API для вишлистов: авторизация, CRUD вишлистов/подарков, публичный
    просмотр и резервирование.'
  title: Wishlist Service API
  version: "1.0"
paths:
...
  /wishlists/{wishlistID}/items:
    get:
      description: Возвращает позиции в указанном вишлисте текущего пользователя.
      parameters:
      - description: Wishlist ID (UUID)
        in: path
        name: wishlistID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.WishlistItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - BearerAuth: []
      summary: List wishlist items
      tags:
      - WishlistItems
    post:
      consumes:
      - application/json
      description: Создает позицию в указанном вишлисте текущего пользователя.
      parameters:
      - description: Wishlist ID (UUID)
        in: path
        name: wishlistID
        required: true
        type: string
      - description: Wishlist item create payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateWishlistItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.WishlistItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - BearerAuth: []
      summary: Create wishlist item
      tags:
      - WishlistItems
  /wishlists/{wishlistID}/items/{itemID}:
    delete:
      description: Удаляет позицию из указанного вишлиста текущего пользователя.
      parameters:
      - description: Wishlist ID (UUID)
        in: path
        name: wishlistID
        required: true
        type: string
      - description: Item ID (UUID)
        in: path
        name: itemID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - BearerAuth: []
      summary: Delete wishlist item
      tags:
      - WishlistItems
    get:
      description: Возвращает позицию по ID из указанного вишлиста текущего пользователя.
      parameters:
      - description: Wishlist ID (UUID)
        in: path
        name: wishlistID
        required: true
        type: string
      - description: Item ID (UUID)
        in: path
        name: itemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.WishlistItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - BearerAuth: []
      summary: Get wishlist item by ID
      tags:
      - WishlistItems
    patch:
      consumes:
      - application/json
      description: Частично обновляет позицию в указанном вишлисте текущего пользователя.
      parameters:
      - description: Wishlist ID (UUID)
        in: path
        name: wishlistID
        required: true
        type: string
      - description: Item ID (UUID)
        in: path
        name: itemID
        required: true
        type: string
      - description: Wishlist item patch payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateWishlistItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.WishlistItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - BearerAuth: []
      summary: Update wishlist item
      tags:
      - WishlistItems
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'JWT access token with Bearer prefix. Example: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"